\name{format.dms}
\alias{format.dms}
\title{ Format an angle for printing }
\description{
  Function for pretty formatting (degrees, minutes, seconds) for angular
  class of data.
}
\usage{
\method{format}{dms}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ object to format }
  \item{\dots}{ additional arguments }
}
\value{
  String with formatted data.
}
\author{ Lukasz Komsta }
\keyword{ manip }

