\name{format.jd}
\alias{format.jd}
\title{ Format Julian Day Number for pretty printing }
\description{
  Convert Julian Day Number back to date and create string with formatted output.
}
\usage{
\method{format}{jd}(x, ...)
}
\arguments{
  \item{x}{ object to format }
  \item{\dots}{ additional arguments }
}
\author{ Lukasz Komsta }
\keyword{ manip }

