\name{format.time}
\alias{format.time}
\title{ Format time-related data for pretty printing }
\description{
  Format data expressed as hours from midnight to hours, minutes and seconds.
}
\usage{
\method{format}{time}(x, ...)
}
\arguments{
  \item{x}{ object to format }
  \item{\dots}{ additional arguments }
}
\author{ Lukasz Komsta }
\keyword{ manip }

