\name{moonsun-package}
\alias{moonsun-package}
\alias{moonsun}
\docType{package}
\title{
Basic astronomical routines with R
}
\description{
A collection of basic astronomical routines
for R - coordinates conversion, ephemerides, rise and
set computing, prediction of angles between objects and
so on.
}
\details{
\tabular{ll}{
Package: \tab moonsun\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2008-09-03\cr
License: \tab GPL2\cr
}
}
\author{
Maintainer: Lukasz Komsta <luke@novum.am.lublin.pl>
}
\references{
Duffet-Smith, P. Practical Astronomy with your calculator, Third Edition. Cambridge University Press, 2004.
}
\keyword{ package }

