\name{print.eqc}
\alias{print.eqc}
\alias{print.jd}
\alias{print.time}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Format astronomical data for pretty printing }
\description{
  Internal functions for pretty printing.
}
\usage{
\method{print}{eqc}(x, ...)
\method{print}{jd}(x, ...)
\method{print}{time}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ object to be formatted }
  \item{\dots}{ another arguments }
}
\author{ Lukasz Komsta }
\keyword{ print }
